Separate T3DCreate* into its own header to avoid including Windows.h etc in some cases.
