+++ android/jni/AndroidVulkanContext.cpp
@@ -6,6 +6,7 @@
 #include "Common/Vulkan/VulkanLoader.h"
 #include "Common/Vulkan/VulkanContext.h"
 #include "thin3d/VulkanRenderManager.h"
+#include "thin3d/thin3d_create.h"
 #include "util/text/parsers.h"
 #include "Core/Config.h"
 #include "Core/System.h"
