+++ ext/native/thin3d/thin3d.h
@@ -6,7 +6,7 @@
 
 #pragma once
 
-#include <stdint.h>
+#include <cstdint>
 #include <cstddef>
 #include <vector>
 #include <string>
@@ -15,26 +15,6 @@
 #include "DataFormat.h"
 
 class Matrix4x4;
-
-#ifdef _WIN32
-#ifndef NOMINMAX
-#define NOMINMAX
-#endif
-#define WIN32_LEAN_AND_MEAN
-#include <Windows.h>
-#include <d3dcommon.h>
-struct IDirect3DDevice9;
-struct IDirect3D9;
-struct IDirect3DDevice9Ex;
-struct IDirect3D9Ex;
-struct ID3D11Device;
-struct ID3D11DeviceContext;
-struct ID3D11Device1;
-struct ID3D11DeviceContext1;
-
-#endif
-
-class VulkanContext;
 
 namespace Draw {
 
@@ -664,16 +644,7 @@ protected:
 	int targetHeight_;
 };
 
-DrawContext *T3DCreateGLContext();
-
 extern const UniformBufferDesc UBPresetDesc;
-
-#ifdef _WIN32
-DrawContext *T3DCreateDX9Context(IDirect3D9 *d3d, IDirect3D9Ex *d3dEx, int adapterId, IDirect3DDevice9 *device, IDirect3DDevice9Ex *deviceEx);
-DrawContext *T3DCreateD3D11Context(ID3D11Device *device, ID3D11DeviceContext *context, ID3D11Device1 *device1, ID3D11DeviceContext1 *context1, D3D_FEATURE_LEVEL featureLevel, HWND hWnd);
-#endif
-
-DrawContext *T3DCreateVulkanContext(VulkanContext *context, bool split);
 
 // UBs for the preset shaders
